// src/trading/models/OrderSide.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trading API
 * Alpaca\'s Trading API is a modern platform for algorithmic trading.
 *
 * The version of the OpenAPI document: 2.0.1
 * Contact: support@alpaca.markets
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * Represents which side this order was on:
 * - buy
 * - sell
 * Required for all order classes except for mleg.
 * @export
 */
export const OrderSide = {
    Buy: 'buy',
    Sell: 'sell'
} as const;
export type OrderSide = typeof OrderSide[keyof typeof OrderSide];


export function instanceOfOrderSide(value: any): boolean {
    for (const key in OrderSide) {
        if (Object.prototype.hasOwnProperty.call(OrderSide, key)) {
            if (OrderSide[key as keyof typeof OrderSide] === value) {
                return true;
            }
        }
    }
    return false;
}

export function OrderSideFromJSON(json: any): OrderSide {
    return OrderSideFromJSONTyped(json, false);
}

export function OrderSideFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderSide {
    return json as OrderSide;
}

export function OrderSideToJSON(value?: OrderSide | null): any {
    return value as any;
}

export function OrderSideToJSONTyped(value: any, ignoreDiscriminator: boolean): OrderSide {
    return value as OrderSide;
}