// src/trading/models/OrderClass.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trading API
 * Alpaca\'s Trading API is a modern platform for algorithmic trading.
 *
 * The version of the OpenAPI document: 2.0.1
 * Contact: support@alpaca.markets
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


/**
 * The order classes supported by Alpaca vary based on the order's security type. The following provides a comprehensive breakdown of the supported order classes for each category:
 *   - Equity trading: simple (or ""), oco, oto, bracket.
 *   - Options trading:
 *     - simple (or "")
 *     - mleg (required for multi-leg complex option strategies)
 *   - Crypto trading: simple (or "").
 * @export
 */
export const OrderClass = {
    Simple: 'simple',
    Bracket: 'bracket',
    Oco: 'oco',
    Oto: 'oto',
    Mleg: 'mleg',
    Empty: ''
} as const;
export type OrderClass = typeof OrderClass[keyof typeof OrderClass];


export function instanceOfOrderClass(value: any): boolean {
    for (const key in OrderClass) {
        if (Object.prototype.hasOwnProperty.call(OrderClass, key)) {
            if (OrderClass[key as keyof typeof OrderClass] === value) {
                return true;
            }
        }
    }
    return false;
}

export function OrderClassFromJSON(json: any): OrderClass {
    return OrderClassFromJSONTyped(json, false);
}

export function OrderClassFromJSONTyped(json: any, ignoreDiscriminator: boolean): OrderClass {
    return json as OrderClass;
}

export function OrderClassToJSON(value?: OrderClass | null): any {
    return value as any;
}

export function OrderClassToJSONTyped(value: any, ignoreDiscriminator: boolean): OrderClass {
    return value as OrderClass;
}