// src/trading/models/PostOrderRequest.ts
export interface PostOrderRequest {
    /**
     * 
     * @type {AdvancedInstructions}
     * @memberof PostOrderRequest
     */
    advancedInstructions?: AdvancedInstructions;
    /**
     * A unique identifier for the order. Automatically generated if not sent. (<= 128 characters)
     * @type {string}
     * @memberof PostOrderRequest
     */
    clientOrderId?: string;
    /**
     * (default) false. If true, order will be eligible for execution in the pre-market, after-hours, and overnight sessions. Only works with type `limit` and time_in_force set to either `day` or `gtc`
     * @type {boolean}
     * @memberof PostOrderRequest
     */
    extendedHours?: boolean;
    /**
     * list of order legs (<= 4)
     * @type {Array<MLegOrderLeg>}
     * @memberof PostOrderRequest
     */
    legs?: Array<MLegOrderLeg>;
    /**
     * Required if type is `limit` or `stop_limit`.
     * In case of `mleg`, the limit_price parameter is expressed with the following notation:
     * - A positive value indicates a debit, representing a cost or payment to be made.
     * - A negative value signifies a credit, reflecting an amount to be received.
     * @type {string}
     * @memberof PostOrderRequest
     */
    limitPrice?: string;
    /**
     * dollar amount to trade. Cannot work with `qty`. Can only work for market order types and day for time in force.
     * @type {string}
     * @memberof PostOrderRequest
     */
    notional?: string;
    /**
     * 
     * @type {OrderClass}
     * @memberof PostOrderRequest
     */
    orderClass?: OrderClass;
    /**
     * 
     * @type {PositionIntent}
     * @memberof PostOrderRequest
     */
    positionIntent?: PositionIntent;
    /**
     * number of shares to trade. Can be fractionable for only market and day order types. Required for `mleg` order class, represents the number of units to trade of this strategy.
     * @type {string}
     * @memberof PostOrderRequest
     */
    qty?: string;
    /**
     * 
     * @type {OrderSide}
     * @memberof PostOrderRequest
     */
    side?: OrderSide;
    /**
     * 
     * @type {PostOrderRequestStopLoss}
     * @memberof PostOrderRequest
     */
    stopLoss?: PostOrderRequestStopLoss;
    /**
     * required if type is `stop` or `stop_limit`
     * @type {string}
     * @memberof PostOrderRequest
     */
    stopPrice?: string;
    /**
     * symbol, asset ID, or currency pair to identify the asset to trade, required for all order classes except for `mleg`.
     * @type {string}
     * @memberof PostOrderRequest
     */
    symbol?: string;
    /**
     * 
     * @type {PostOrderRequestTakeProfit}
     * @memberof PostOrderRequest
     */
    takeProfit?: PostOrderRequestTakeProfit;
    /**
     * 
     * @type {TimeInForce}
     * @memberof PostOrderRequest
     */
    timeInForce: TimeInForce;
    /**
     * this or `trail_price` is required if type is `trailing_stop`
     * @type {string}
     * @memberof PostOrderRequest
     */
    trailPercent?: string;
    /**
     * this or `trail_percent` is required if type is `trailing_stop`
     * @type {string}
     * @memberof PostOrderRequest
     */
    trailPrice?: string;
    /**
     * 
     * @type {OrderType}
     * @memberof PostOrderRequest
     */
    type: OrderType;
}