// src/trading/models/PostOrderRequestTakeProfit.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trading API
 * Alpaca\'s Trading API is a modern platform for algorithmic trading.
 *
 * The version of the OpenAPI document: 2.0.1
 * Contact: support@alpaca.markets
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Takes in a string/number value for limit_price
 * @export
 * @interface PostOrderRequestTakeProfit
 */
export interface PostOrderRequestTakeProfit {
    /**
     * 
     * @type {string}
     * @memberof PostOrderRequestTakeProfit
     */
    limitPrice?: string;
}

/**
 * Check if a given object implements the PostOrderRequestTakeProfit interface.
 */
export function instanceOfPostOrderRequestTakeProfit(value: object): value is PostOrderRequestTakeProfit {
    return true;
}

export function PostOrderRequestTakeProfitFromJSON(json: any): PostOrderRequestTakeProfit {
    return PostOrderRequestTakeProfitFromJSONTyped(json, false);
}

export function PostOrderRequestTakeProfitFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostOrderRequestTakeProfit {
    if (json == null) {
        return json;
    }
    return {
        
        'limitPrice': json['limit_price'] == null ? undefined : json['limit_price'],
    };
}

export function PostOrderRequestTakeProfitToJSON(json: any): PostOrderRequestTakeProfit {
    return PostOrderRequestTakeProfitToJSONTyped(json, false);
}

export function PostOrderRequestTakeProfitToJSONTyped(value?: PostOrderRequestTakeProfit | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'limit_price': value['limitPrice'],
    };
}