// src/trading/models/PostOrderRequestStopLoss.ts
/* tslint:disable */
/* eslint-disable */
/**
 * Trading API
 * Alpaca\'s Trading API is a modern platform for algorithmic trading.
 *
 * The version of the OpenAPI document: 2.0.1
 * Contact: support@alpaca.markets
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */

import { mapValues } from '../runtime';
/**
 * Takes in string/number values for stop_price and limit_price
 * @export
 * @interface PostOrderRequestStopLoss
 */
export interface PostOrderRequestStopLoss {
    /**
     * 
     * @type {string}
     * @memberof PostOrderRequestStopLoss
     */
    limitPrice?: string;
    /**
     * 
     * @type {string}
     * @memberof PostOrderRequestStopLoss
     */
    stopPrice?: string;
}

/**
 * Check if a given object implements the PostOrderRequestStopLoss interface.
 */
export function instanceOfPostOrderRequestStopLoss(value: object): value is PostOrderRequestStopLoss {
    return true;
}

export function PostOrderRequestStopLossFromJSON(json: any): PostOrderRequestStopLoss {
    return PostOrderRequestStopLossFromJSONTyped(json, false);
}

export function PostOrderRequestStopLossFromJSONTyped(json: any, ignoreDiscriminator: boolean): PostOrderRequestStopLoss {
    if (json == null) {
        return json;
    }
    return {
        
        'limitPrice': json['limit_price'] == null ? undefined : json['limit_price'],
        'stopPrice': json['stop_price'] == null ? undefined : json['stop_price'],
    };
}

export function PostOrderRequestStopLossToJSON(json: any): PostOrderRequestStopLoss {
    return PostOrderRequestStopLossToJSONTyped(json, false);
}

export function PostOrderRequestStopLossToJSONTyped(value?: PostOrderRequestStopLoss | null, ignoreDiscriminator: boolean = false): any {
    if (value == null) {
        return value;
    }

    return {
        
        'limit_price': value['limitPrice'],
        'stop_price': value['stopPrice'],
    };
}